library/requirements.txt
asgiref==3.3.1
Django==1.11.6
djangorestframework==3.9.4
Pillow==8.1.0
pkg-resources==0.0.0
pytz==2020.5
sqlparse==0.4.1
